refactor: rest icon components to tsx
