+++ src/components/IconText.tsx
@@ -1,7 +1,12 @@
 import { Paragraph, XStack } from 'tamagui'
-import { Icon } from '../Icon/Icon'
+import Icon from './Icon'
 
-const IconText = ({ icon, text, ...props }) => {
+type IconTextProps = {
+  icon: string
+  children: string
+}
+
+const IconText = ({ icon, children, ...props }: IconTextProps) => {
   return (
     <XStack
       style={{
@@ -11,7 +16,7 @@ const IconText = ({ icon, text, ...props }) => {
     >
       <Icon source={icon} width={16} height={16} />
       <Paragraph {...props} color={'#000000'}>
-        {text}
+        {children}
       </Paragraph>
     </XStack>
   )
